Change secret schema type to avoid no password prompt
As shown in #213, mailnag won't authenticate on first start. The reason
can be found in https://gitlab.gnome.org/GNOME/libsecret/-/issues/7. It
is because that when creating schema using `Secret.SchemaFlags.NONE`.
The schema name is sent as attribute to match (`xdg:schema`). However,
for gnome-keyring side, when the secret is locked. The secret will have
extra attributes which makes that the lookup fails. And use
`Secret.Schema.Flags.DONT_MATCH_NAME` flag creates the schema can avoid
this situation.

However, there is also another situation, just make the schema name to
be `None`, which would be `org.freedesktop.Secret.Generic` schema, which
means no schema at all, so any attributes can be used. So the
`password_lookpup_sync` can also success to allow the password prompt.

This PR adopts the first solution, simply change the FLAG.

Closes #213